feat(idempotency): implement idempotency key mechanism with persistence layer
- Add IdempotencyKey entity and repository
- Implement IdempotencyService with request deduplication logic
- Ensure safe execution of idempotent operations
- Integrate with order operations for reliability

+++ src/main/java/com/logibridge/backend/order/dto/OrderResponse.java
@@ -23,5 +23,5 @@ public class OrderResponse {
     private OrderStatus status;
     private LocalDateTime createdAt;
 
-    private Long trackingCount;
+    private int trackingCount;
 }
